requirements.txt

fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
websockets==12.0
pydantic==2.8.2

transformers==4.42.4
torch>=2.1
torchaudio>=2.1
sentencepiece
protobuf

openai-whisper==20231117
faster-whisper==1.0.3

pydub==0.25.1
numpy
scikit-learn
librosa==0.10.2.post1
soundfile

coqpit
TTS==0.22.0

moviepy==1.0.3
